% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAge.R
\name{getAge}
\alias{getAge}
\title{Get Age Function}
\usage{
getAge(RawDataSet)
}
\arguments{
\item{RawDataSet}{The name of your raw data set that contains the desired items.}
}
\description{
This function returns a  dataframe with the Age (in years) of each participant
}
\examples{
getAge()
}
\keyword{Age,}
\keyword{WiLD}
\keyword{Years,}
