% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSex.R
\name{getSex}
\alias{getSex}
\title{Get Sex Function}
\usage{
getSex(RawDataSet)
}
\arguments{
\item{RawDataSet}{The name of your raw data set that contains the desired items.}
}
\description{
This function returns a factored dataframe with the Sex of each participant
}
\examples{
getSex()
}
\keyword{Female}
\keyword{Gender,}
\keyword{Male,}
\keyword{Sex,}
