% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profPersonality.R
\name{profPersonality}
\alias{profPersonality}
\title{Profile Personality Function}
\usage{
profPersonality(RawDataSet)
}
\arguments{
\item{RawDataSet}{The name of your raw data set that contains the desired items.}
}
\description{
This function returns a dataframe with the personality scores for each participant
}
\examples{
profPersonality()
}
\keyword{Big}
\keyword{Five,}
\keyword{IPIP}
\keyword{Personality,}
\keyword{Profile,}
