% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullWiLD.R
\name{pullWiLD}
\alias{pullWiLD}
\title{pullWiLD Function}
\usage{
pullWiLD(RawData, tools)
}
\arguments{
\item{RawData}{A dataframe of the raw data}

\item{tools}{A character vector of tool acronyms (e.g., LUPI, CAPI, GGD)}
}
\description{
This function returns data from the specified tools in the WiLD Toolkit.
}
\examples{
##Returns CAPI and LUPI Data
pullWiLD(RawData, c("LUPI", "CAPI"))

##Returns Goals Going Deeper
pullWiLD(RawData, "GGD")
}
\keyword{Toolkit,}
\keyword{WiLD,}
\keyword{data,}
\keyword{subset}
